Updated build to properly set context configuration
I'd forgotten that I'd hacked this into PGX::Build last year. The default
configuration is "dev", and the default configuration file is thus
`conf/dev.yml`. I added this to `VersionCompare.pm`, and now `./Build` will
change that value as appropriate in `blib/lib/PGX/VersionCompare.pm`. So if
you run `./Build --context stage`, it will change
`blib/lib/PGX/VersionCompare.pm` to use `conf/stage.yml`. So then all I had to
do was update the various starupt scripts in the `scripts` directory to look
for modules in `blib/lib` before looking in `lib`. The cool thing is that,
although the default context is "test", the default configuration is "dev".
This means that you can build for running regression tests on the test
database, but when you run the server, it will use the dev database for the
UI.

I took the opportunity to update the `name:` setting in each of the
configuration files to properly reflect the context, so you can see the name
of the context in the title bar of the app when you run it.