Got basic URL-based functionality working.  Needs work, tests before pushing.
Things that work
******************************************************************************

- URL syntax:  Now you can go to http://localhost:3000/version/8.1.0/8.1.4
and it will show the form (with sticky fields) and the results of your
query.

Things that don't work
******************************************************************************

- JavaScript is missing.  We need it to constrain the values of the
minor versions select box to only those versions that actually exist within
the major version that the user selected.  For bootstrapping purposes,
I cheated and hard-coded 0-20 as options for the minor versions select box.

- Submission via the Show button isn't written.  The most elegant way to do
this would be to use JS, on form submission, to set the action URL dynamically.

- Search term is ignored. You can do http://localhost:3000/version/8.1.0/8.1.4 ,
but if you do http://localhost:3000/version/8.1.0/8.1.4?q=Hiroshi
you get the same thing (when you should get only fixes that contain "Hiroshi.")
Unlike the forgoing, this functionality is coded, so there must be a bug.
I'll fix it in the next commit.

- CSS is missing.

- Controller and view tests are missing.

- POD coverage is incomplete.